[Issue #7353] Fix error when loading data from tcertificates table (#7354)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #7353 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Added `tcertificates.is_selfsigned` to `columns_to_exclude` in the
LoadOracle task. Updated the `tcertificates mixin` to allow
`is_selfsigned` to be `None`.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
The `is_selfsigned` field was basically throwing an error when we were
trying to copy it. By removing it, and also allowing that field to be
`None`, we should just prevent that issue from occuring.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
Manually run the command and the `staging.tcertificates` table should
populate.