Enable API Gateway auth enforcement (#6764)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #ISSUE 

## Changes proposed

- Added missing endpoints that require no API gateway auth
- Enforce auth on full proxy path

## Context for reviewers

This PR will be the last change from the API Gateway side for forcing
authentication at the gateway. We have defined endpoints that require no
auth because they are either used for authenticating, are swagger
related, are related to bot/DNS, or have a custom downstream auth
(legacy SOAP), and the rest are enforcing authentication.

### 🔴 Manual deploy steps

From testing, I had to manually delete the existing endpoints from the
resources (which doesn't impact anything until it's deployed) since
terraform output an error saying the resources already existed. So, this
means you MUST do this, or the deploy will fail

1. Go to the API Gateway console
2. Select the gateway that is being updated
3. If it doesn't put you on the resources tab of the page, click on that
4. Delete all resources
    - Again, this is required or the infra will not deploy
- This will NOT impact any existing traffic using the gateway, it will
just be deployed from the new terraform once the depends_on are created
    - Delete these V
<img width="321" height="613" alt="Screenshot 2025-12-04 at 1 29 03 PM"
src="https://github.com/user-attachments/assets/7bb7f1d5-e213-4614-8ac0-8c74a3684636"
/>

## Validation steps

Brandon has a few scripts used to hit the various endpoints to verify
things are working as expected, and it passed as of 12/4!