[Issue #6942] adding the skeleton page for manage users legacy invite (#7346)
## Summary

Work for #6942 

## Changes proposed

Building out a skeleton for the Invite legacy page that shows the
organization information and the empty use case, plus the error use case
if we can't load the data.

## Context for reviewers

This is the first ticket out of 3 for building out this page, the
authorized data/table data will be added in the [next
ticket](https://github.com/HHS/simpler-grants-gov/issues/7247).

A word on unittests: We can't currently unittest pages that are behind
feature flags. Working with @doug-s-nava in a different PR to try to
sort that. So in the short term I don't have unittests for that, I'll
include them in the follow on PR with the table.

I also wanted to add a unittest for adding the button to the manage
users page, but the unittests in ManageUsersPageContent.test.tsx only
check calls to mocks not the rendered output, and apparently there is a
problem there which renders and empty div. I didn't have time to debug
that, so I couldn't add a test.

I also slipped in switching the view details button and the manage users
button on the organizations page, that's so it looks nicer when a user
is sometimes a member and sometimes an admin the UI will stay stable.

## Validation steps

- [ ] Make sure that the `manageUsersOff` feature flag is disabled
- [ ] Log on as a user who is an admin of an organization (e.g.
one_org_user on localhost)
- [ ] Go through the Header: Workspace -> Organizations
- [ ] Click on "Manage users" for an org
- [ ] Click on the button saying "Invite Grants.gov users"
- [ ] Note that the breadcrumbs are visible and the links work as
expected
- [ ] Note the org title is listed in the header
- [ ] See the message implying that the request for data is empty and
that the link works correctly
- [ ] Save the URL, log out, and then attempt to go to the URL again. It
should take you to the page notifiying you to log in
- [ ] To see the error case, open `organization_routes.py`, include the
import `from src.api.route_utils import raise_flask_error` at the top,
and add `raise_flask_error(500, f"temp error!!")` at line 60
- [ ] Note the page displays with limited data (no org name, change in
breadcrumbs) and a polite error message
- [ ] Check the organizations page that the manage users button and view
detail button are in that order (the manage users button will not be
present if they are only the member of an org)