Fix bug for created_at and updated_at in common grants search (#7396)
## Summary


Fixes [bug reported in common-grants sdk
](https://github.com/HHS/simpler-grants-gov/issues/7290)

## Changes proposed

Fixed mapping for `updated_at` and `created_at` fields in the service
for the common-grants wrapper APIs. Instead of using the root level
object we are updating the transform to use the dates from the summary
object field.

## Context for reviewers

This change only impacts the common-grants specific code and leaves
untouched the v1 routes.

## Validation steps

Call either the common grants search route using curl or postman and
validate that the `updated_at` and `created_at` dates are populating
correctly. Most of these dates should not be todays date, which is what
was happening prior to this code change