[Issue #7083] Update `list-forms` script to include commands for updating form instructions (#7422)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #7083  

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
In `list_forms_task`, if a form has a `form_instruction_id`, also
include the command to update the instruction.
Update tests.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
We want to update our ListFormTask to help generate the command to run
the new task created in #7082. The list form task as part of checking
whether forms are up-to-date, if any are not up-to-date will
conveniently print out commands saying how to update them.

We want to update this task to also print out commands for how to update
the form instructions. We'll want to build a similar command string, but
for this new script. As that script hasn't yet been written, exact field
names aren't yet known, but the command it generates should look roughly
like:

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
See updated tests and expected output from task.