[Issue #7089] Add FK from Opportunity table to Agency table (#7409)
## Summary

Fixes #7089 

## Changes proposed

We want to add an agency_id column to the Opportunity table, this would
be a nullable foreign key to the Agency table.

## Context for reviewers

Right now we have a hacky approach to getting the Agency connection by
joining on the agency_code string, we want to instead work towards
having a proper foreign key and no longer relying on this hacky
workaround.

We'll need to handle actually populating this column separately, we just
want the column added, and follow-ups will handle populating the value.

## Validation steps

- agency_id foreign key added to the Opportunity table
- DB Migration created

---------

Co-authored-by: Hao Wang <hao.wang@reisystems.com>