Add dockleconfig "settings.py" ignore configuration for NOFO Builder (#7442)
## Summary

Dockle wrongly flags these `settings.py` files these as some kind of
sensitive file, but it is a well-known Python convention.

By adding these filepaths to the `.dockleconfig` file, we are telling
Dockle to ignore these files.

The previous configuration we were using:

- Was deleted in this commit:
https://github.com/HHS/simpler-grants-gov/commit/72ccbf2266e0be97a7783e665b833d94d091c737
- Needed to be updated anyway since we upgraded to Python 3.14.2:
https://github.com/HHS/simpler-grants-pdf-builder/pull/575

Once this is in, it fixes automated deployments for the NOFO Builder.