Update MAINTAINERS and CODEOWNERS to add @kkrug (#7448)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #7443  
Add @kkrug to MAINTAINERS.md and CODEOWNERS

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
MAINTAINERS:
- 'Kenny Krug' added under 'Engnineering'

CODEOWNERS:
- @kkrug added to /documentation/
- @kkrug added to /api/
- @kkrug added to /documentation/api/

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
Confirm @kkrug is listed in the appropriate places in both files