[Issue #7350] Change job to generate Zip of successfully submitted application to run hourly (#7506)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #7350

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Update Terraform to run `create-application-submission` task hourly.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
This was updated to offset by 15 minutes to avoid collision with other
hourly jobs, but this can be changed back to on-the-hour.

**Do not merge until
https://github.com/HHS/simpler-grants-gov/issues/7350 is closed.**

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
Once merged, confirm TF applies these to environments.