[Unticketed] Fix pagination logic to return the item not raise (#7546)
## Summary

## Changes proposed
* Fixes an issue with a recent change which should return the invalid
value rather than raise it

## Context for reviewers
This function is called for pre-processing pagination parameters in a
request. If a user passes in `{"pagination": 5}` instead of the proper
pagination object, this first if-statement prevents us from parsing the
value and returns it. Other type validation will then kick in and say
"that's not an object" and error appropriately. We mistakenly were
raising the object as an error, and `raise 5` has no meaning in python
and gives a separate error entirely.

## Validation steps
Tests updated, but found by manually adjusting params locally when
calling endpoint as well.