[Issue #7394] setup simpler multiauth for JWT and new API key (#7523)
## Summary

Fixes #7394  

## Changes proposed

- Create a
[multiauth](https://github.com/HHS/simpler-grants-gov/blob/main/api/src/auth/multi_auth.py)
class alternative to MultiHttpTokenAuth - it should have a get_user
function that returns an actual user - not one of the auth types like
the existing one (we don't need the key, we want the user itself). The
only possible auth types it should support are UserApiKey and
UserTokenSession
- Define an auth object in the file that allows for either of the two
auths
- Add tests that verify this auth can be used on a generic route (don't
add it to a real route yet)
- Add docs on how to use it - calling out that you can't use the
APIFlask decorator due to it not supporting multiauth yet

## Context for reviewers

For our endpoints, we can configure multiple auth types to be allowed,
we want to add JWT and API User Key auth to many of our endpoints, but
first need to setup the multiauth to support this.

## Validation steps

- New multiauth container class created which has a get_user method that
always returns a user record regardless of which of the two auths it is.
- Tests added that verify the auth works

---------

Co-authored-by: Hao Wang <hao.wang@reisystems.com>