[Issue #7058] allow frontend to process schema with complex nested conditionals (#7482)
The [mergeAllOf](https://github.com/mokkabonna/json-schema-merge-allof)
library the frontend uses to manage processing "allOf" nodes in the form
Json schema is not able to deal with nodes that represent complex
conditional logic. As a result, forms that rely on this logic will not
load in the frontend. (See the NEH cover sheet form here as an example
https://github.com/HHS/simpler-grants-gov/pull/7056/files#diff-fb95c16498447c8cef4c653f086f75b126254e9fc79a774835ab502ceeed60ecR282)

This change removes any relevant allOf nodes from the form json schema,
and returns them as a separate list of conditional validations to be
referenced later on (though this functionality is currently unused).