[Issue #7452] Attachments failing with API Gateway (#7486)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Work for #7452

## Changes proposed

API Key was missing on request causing API Gateway to reject the
request. That rejection was getting eaten and no details about what was
actually happening was getting logged or returned to the FE. The
requests were never reaching the Python API as a result.

## Context for reviewers

After discovering the initial cause, we made a few additional
improvements to how the Next API layer was calling the Python API in
this case.

## Validation steps 
(This testing has already been validated locally on the branch by Doug
and Matt, and in Dev with this branch deployed by Matt. So this list is
purely for UAT, reviewers don't have to take these steps locally on this
branch)

1. Log in
1. Go to an existing Application or start a new Application
1. Upload a file to the Attachments section at the bottom of the
Application page
1. Validate that the file uploads successfully (doesn't get stuck on the
waiting spinner)
1. Go into a form that requires an attachment, like the budget narrative
attachement form and upload a file there.
1. Validate that the file uploads successfully (doesn't get stuck on the
waiting spinner)
1. Save the form
1. Refresh both pages to validate the files still show up as appropriate

---------

Co-authored-by: Doug Schrashun <dougschrashun@navapbc.com>