[Issue #7641] Enable xml generaion dev and staging envs (#7650)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes #7641

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Enable `ENABLE_XML_GENERATION` in dev + staging environments.

## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
XML was tested as generating in staging and the SF-424 and associated
task look good. One AC for 7641 is to enable in dev/staging. which is
accomplished by this PR.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
Confirm ENV var is enabled in dev/staging after deploy.