[Issue #6663] Fetch simpler data for get submission list expanded (#7525)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes / Work for #6663 

## Changes proposed

<!-- What was added, updated, or removed in this PR. -->
Wiring up GetSubmissionListExpanded which means putting it into the
SimplerGrantorsS2SClient and add tests that prove that the client can
handle a request. The grantors services part had to be updated to
produce a dictionary correctly to be turned into xml.



## Context for reviewers

<!-- Technical or background context, more in-depth details of the
implementation, and anything else you'd like reviewers to know about
that will help them understand the changes in the PR. -->
Basically, a request to GetSubmissionListExpanded should get an xml
response based on submissions from our db. Currently this does not
happen because GetSubmissionListExpanded is not an operation that
exists. This PR should add that operation, and then it should be
something that produces data from our db and accepts the filters from
the legacy endpoint.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->