[Issue #7495] Add organization added event on application creation (#7714)
## Summary

<!-- Use "Fixes" to automatically close issue upon PR merge. Use "Work
for" when UAT is required. -->
Fixes for #7495  
Ensure the correct audit events are published when an application is
created with an organization owner.
## Changes proposed
Added logic to publish an ORGANIZATION_ADDED audit event when an
application is created and owned by organization
<!-- What was added, updated, or removed in this PR. -->

## Context for reviewers
Per the issue requirements, when an application is started as being
owned by an organization, two audit events should be published:
`APPLICATION_CREATED`
`ORGANIZATION_ADDED`
Previously, only the application created event was emitted.

## Validation steps

<!-- Manual testing instructions, as well as any helpful references
(screenshots, GIF demos, code examples or output). -->
Create an application owned by an organization 
Verify that  both events exist in the application history page